+++ Runtime/Framework/State/IState.cs
@@ -6,7 +6,7 @@ using UnityEngine.Events;
 
 namespace DT.UniStart {
   #region State
-  public interface IState<T> : IWatchable, IWatchable<T>, IWatchable<T, T>, IGetValue<T> { }
+  public interface IState<T> : IWatchable<T, T>, IGetValue<T> { }
   public interface IListState<T> : IReadOnlyList<T>, IWatchable, IWatchable<ReadOnlyCollection<T>>, IGetValue<ReadOnlyCollection<T>> {
     bool Contains(T item);
     int IndexOf(T item);
@@ -30,6 +30,8 @@ namespace DT.UniStart {
     public static IListState<T> AddList<T>(this IStateManager manager, List<T> value) => new WatchList<T>(value);
     public static IDictionaryState<K, V> AddDictionary<K, V>(this IStateManager manager) => new WatchDictionary<K, V>();
     public static IDictionaryState<K, V> AddDictionary<K, V>(this IStateManager manager, Dictionary<K, V> value) => new WatchDictionary<K, V>(value);
+
+    #region Echoed
     public static IState<T> Add<T>(this IStateManager manager, out Watch<T> echoed, T value) {
       var state = new Watch<T>(value);
       echoed = state;
@@ -65,12 +67,6 @@ namespace DT.UniStart {
       echoed = state;
       return state;
     }
-    #region Reversed
-    public static IState<T> Add<T>(this IStateManager manager, T value, out Watch<T> echoed) => manager.Add(out echoed, value);
-    public static IListState<T> AddArray<T>(this IStateManager manager, int count, out WatchArray<T> echoed) => manager.AddArray(out echoed, count);
-    public static IListState<T> AddArray<T>(this IStateManager manager, T[] value, out WatchArray<T> echoed) => manager.AddArray(out echoed, value);
-    public static IListState<T> AddList<T>(this IStateManager manager, List<T> value, out WatchList<T> echoed) => manager.AddList(out echoed, value);
-    public static IDictionaryState<K, V> AddDictionary<K, V>(this IStateManager manager, Dictionary<K, V> value, out WatchDictionary<K, V> echoed) => manager.AddDictionary(out echoed, value);
     #endregion
   }
   #endregion